Since this the version number is now stored as a long we'll get a crash unless we try to retrieve it as one. Since older versions of the app stored it as an int, we need to throw the old prefernce away.

+++ app/src/main/java/com/google/android/stardroid/activities/SplashScreenActivity.java
@@ -140,7 +140,7 @@ public class SplashScreenActivity extends InjectableActivity
   }
 
   private void maybeShowWhatsNewAndEnd() {
-    boolean whatsNewSeen = (sharedPreferences.getInt(
+    boolean whatsNewSeen = (sharedPreferences.getLong(
         ApplicationConstants.READ_WHATS_NEW_PREF_VERSION, -1) == app.getVersion());
     if (whatsNewSeen) {
       launchSkyMap();
